+++ Cargo.toml
@@ -16,6 +16,8 @@ members = [
     "clients/sled-agent-client",
     "clients/wicketd-client",
     "cockroach-admin",
+    "cockroach-admin-api",
+    "cockroach-admin-types",
     "common",
     "dev-tools/crdb-seed",
     "dev-tools/omdb",
@@ -110,6 +112,8 @@ default-members = [
     "clients/sled-agent-client",
     "clients/wicketd-client",
     "cockroach-admin",
+    "cockroach-admin-api",
+    "cockroach-admin-types",
     "common",
     "dev-tools/crdb-seed",
     "dev-tools/omdb",
@@ -263,7 +267,9 @@ ciborium = "0.2.2"
 cfg-if = "1.0"
 chrono = { version = "0.4", features = [ "serde" ] }
 clap = { version = "4.5", features = ["cargo", "derive", "env", "wrap_help"] }
+cockroach-admin-api = { path = "cockroach-admin-api" }
 cockroach-admin-client = { path = "clients/cockroach-admin-client" }
+cockroach-admin-types = { path = "cockroach-admin-types" }
 colored = "2.1"
 const_format = "0.2.32"
 cookie = "0.18"
