+++ cockroach-admin-api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cockroach-admin-api"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+cockroach-admin-types.workspace = true
+dropshot.workspace = true
+omicron-common.workspace = true
+omicron-uuid-kinds.workspace = true
+omicron-workspace-hack.workspace = true
+schemars.workspace = true
+serde.workspace = true
