+++ cockroach-admin-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "cockroach-admin-types"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+chrono.workspace = true
+csv.workspace = true
+omicron-common.workspace = true
+omicron-workspace-hack.workspace = true
+schemars.workspace = true
+serde.workspace = true
+
+[dev-dependencies]
+proptest.workspace = true
+test-strategy.workspace = true
