+++ cockroach-admin/Cargo.toml
@@ -12,6 +12,8 @@ anyhow.workspace = true
 camino.workspace = true
 chrono.workspace = true
 clap.workspace = true
+cockroach-admin-api.workspace = true
+cockroach-admin-types.workspace = true
 csv.workspace = true
 dropshot.workspace = true
 http.workspace = true
